Add files via upload
Here, I have focus on working on the Admin side first.
I have tried to create a Login page, the home page, forgot the password, and change the password page.
These pages are working but still need formatting. 

Next week, I am planning to work on the CSS file first so that, I can attach it with my files together for convenience. 
Then, I am planning to work on the Booking page, location page, city page, a code that goes to the registered email or ph no for password recovery, and feedback using an online template for front side view.