Add files via upload
This, I plan to work on Admin and User part. Because there is a lot to do in there, I am just able to focus on the admin part. 
I have added the template and tried to add and remove a lot of stuff from the template according to the requirement of the project. 
Many times, I have to phase the problems with the Xamp and MySQL because of my device. 
There are still a lot of small bugs that I have to fix but for now, I want to move forward with the next step in the project maybe that is the User side or a worker side. I will have a look at it. I will also start focusing on making reports for the project.

I also want to discuss the "Final Project Demo" which is due end of the month but if possible if we start preparing documents for that day then it will be easier that is what I believe.